Refactor code to use forms and add new DTO
The MvcConfig class has been deleted and replaced with MyController for better routing handling. Forms are now used in the admin-page HTML template to change page titles and text, and a new TextDTO class has been added to manage these data. Furthermore, the TITLE and TEXT static fields in EditLandingPageApplication are now dynamically changed through the admin page and displayed on the index page.